refactor GraphQL code
- move to separate directory
- split GraphQL schema definition
